requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
django-haystack==3.2.1
django-taggit==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.4
factory-boy==3.3.0
Faker==21.0.0
gevent==23.9.1
greenlet==3.0.2
idna==3.6
mysqlclient==2.2.1
oauthlib==3.2.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools-scm==8.0.4
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.1.0
zope.event==5.0
zope.interface==6.1
